backend/requirements.txt
aniso8601==8.1.1
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
pkg-resources==0.0.0
protobuf==3.14.0
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
